优化 byrbt 使用table 形式mediainfo

+++ get_desc/info.sh
@@ -3,14 +3,14 @@
 #
 # Author: rachpt@126.com
 # Version: 3.1v
-# Date: 2019-02-28
+# Date: 2019-04-27
 #
 #-------------------------------------#
 # 复制 nfo 文件内容至简介，如果没有 nfo 文件，
 # 则采用 mediainfo 生成主文件的编码信息至临时文件。
 # 自动判断 是否有 nfo 文件，以及 nfo 文件是否下载完成。
 #-------------------------------------#
-
+#
 # 使用 ffmpeg 获取视频缩略图
 gen_screenshots() {
   local step total file screen_file size ratio row column
@@ -34,61 +34,72 @@ gen_screenshots() {
   \rm -f "${ROOT_PATH}/tmp"/thumbnail-[0-9]*.jpg # 通配符，不能使用引号
 
   # 图片上传 img_url_com  img_url_byr
-  upload_image_com "$screen_file"
-  [[ $enable_byrbt == yes ]] && upload_image_byrbt "$screen_file"
-
-  sleep 0.5 && \rm -f "$screen_file"
+  [[ -s "$screen_file" ]] && upload_image_com "$screen_file"  # static.sh
+  [[ $enable_byrbt == yes && -s "$screen_file" ]] && upload_image_byrbt "$screen_file"
+  # delete image if upload successed
+  [[ "$img_url_com" || "$img_url_byr" ]] && sleep 0.5 && \rm -f "$screen_file"
 }
 
 #-------------------------------------#
 # 读取主文件以获得info，提前生成简介将失效
 generate_info_local() {
+  local main_file_dir max_size_file info_gen_desc info_gen_html
+  local _ext_desc="[b]以下是[url=https://github.com/rachpt/AutoSeed] \
+    AutoSeed [/url]自动完成的截图，不喜勿看。[/b]"
+  local _ext_html="<br /><br /><stong>以下是<a \
+    href=\"https://github.com/rachpt/AutoSeed\"> AutoSeed </a>\
+    自动完成的截图，不喜勿看。</strong><br />"
   # 种子文件绝对路径
-  local main_file_dir="${one_TR_Dir}/${one_TR_Name}"
-  debug_func "info:file-dir[$main_file_dir]"  #----debug---
+  main_file_dir="${one_TR_Dir}/${one_TR_Name}"
+  debug_func "info:folder-dir[$main_file_dir]"  #----debug---
   # 使用 mediainfo 生成种子中体积最大文件的 iNFO
   max_size_file="$(\find "$main_file_dir" -type f -exec stat -c "%s %n" {} \;| \
       sort -nr|head -1|sed -E 's/^[0-9 ]+//')"
-  debug_func "info:max-file[$max_size_file]"  #----debug---
-  # 本地简介大小为零
-  if [ ! -s "$source_desc" ]; then
-    local info_generated="$($mediainfo "$max_size_file"| \
-      sed '/Unique/d;/Encoding settings/d;/Complete name/d;/Writing library/d;/Writing application/d')"
+  debug_func "info:max-file-path[$max_size_file]"  #----debug---
+  # 本地简介大小为零，-s 大小不为零，! 取反
+  if [[ ! -s "$source_desc" ]]; then
+    info_gen_desc="$($mediainfo "$max_size_file"|sed -i \
+      '/Unique/d;/Encoding settings/d;/Complete name/ {s%/.*/%%};/Writing library/d;/Writing application/d')"
+    [[ $enable_byrbt == yes ]] && \
+    info_gen_html="$($mediainfo --Output=HTML "$max_size_file"|sed -i \
+      '/Complete name/ {s%/.*/%%};/html>/d;/body>/d;/head>/d;/<META/d')"
   else
-    local info_generated="$(\cat "$source_desc")"
+    info_gen_desc="$(\cat "$source_desc")"
+    [[ $enable_byrbt == yes ]] && \
+    info_gen_html="$(echo "$info_gen_desc"|sed 's/ /\&nbsp; /g;s!$!&<br />!g')"
   fi
   # 缩略图
+  unset img_url_com img_url_byr # clean
   gen_screenshots
   # 存档
-  if [[ $img_url_byr || $img_url_com ]]; then
-    echo -e "$info_generated\n\n[b]以下是[url=https://github.com/rachpt/AutoSeed] AutoSeed [/url]自动完成的截图，不喜勿看。[/b]\n"${max_size_file##*/}"\n[img]$img_url_com[/img]" > "$source_desc"
-    debug_func "info:screens-gened[$img_url_com]"  #----debug---
-    # byrbt bbcode to html
-    [[ $enable_byrbt == yes && $img_url_byr ]] && {
-    echo "$info_generated"|sed 's/ /\&nbsp; /g;s!$!&<br />!g' > "$source_html" 
-    echo "<br /><br /><stong>以下是<a href=\"https://github.com/rachpt/AutoSeed\"> AutoSeed </a>自动完成的截图，不喜勿看。</strong><br />${max_size_file##*/}<br />" >> "$source_html" # 追加至末尾
-    echo "<img src=\"$img_url_byr\" style=\"width: 900px;\" /> <br />" >> "$source_html" # 追加至末尾
-    debug_func "info:screens-byrbt[$img_url_byr]"  #----debug---
-    }
-  else
-    # byrbt bbcode to html
-    [ "$enable_byrbt" = 'yes' ] && [ -s "$source_desc" ] && \
-      sed 's/ /\&nbsp; /g;s!$!&<br />!g' "$source_desc" > "$source_html" 
-  fi
+  [[ "$img_url_com" ]] && \
+    echo -e "${info_gen_desc}\n\n${_ext_desc}\n${max_size_file##*/}
+      \n[img]$img_url_com[/img]" > "$source_desc"
+  # byrbt desc to html
+  [[ $enable_byrbt == yes ]] && \
+    [[ "$info_gen_html" ]] && {
+    echo -e "${info_gen_html}\n${_ext_html}\n<br />${max_size_file##*/}<br />\n \
+      <img src=\"$img_url_byr\" style=\"width: 900px;\" /><br />" > "$source_html" 
+    } || { echo "$info_gen_html" > "$source_html"; }
+  unset img_url_com img_url_byr # clean
 }
 
+#---------------main------------------#
 # 首先判断是否有 nfo 文件，以及nfo是否下载完成
 read_info_file() {
-  if [ ! "$one_TR_Dir" ]; then
+  if [[ ! "$one_TR_Dir" ]]; then
       debug_func "info:one_TR_Dir.0[$one_TR_Dir]"  #----debug---
       one_TR_Dir="$(find "$default_FILE_PATH" -name \
           "$one_TR_Name" 2> /dev/null|head -1)"
       one_TR_Dir="${one_TR_Dir%/*}"
+      debug_func "info:one_TR_Dir.1[$one_TR_Dir]"  #----debug---
+  else
+      debug_func "info:one_TR_Dir[$one_TR_Dir]"  #----debug---
   fi
-  debug_func "info:one_TR_Dir[$one_TR_Dir]"  #----debug---
 
-  if [ "$one_TR_Dir" ]; then
+  if [[ "$one_TR_Dir" ]]; then
     local nfo_file_size nfo_file_path nfo_file_downloaded   
+    [[ $one_TR_Dir =~ .*/$ ]] && one_TR_Dir=${one_TR_Dir%/} # move slash end
     nfo_file_size=$("$tr_show" "$torrent_Path"| \
       grep -Eio '\.nfo \([0-9\. ]+[kb]+\)'|grep -Eo '[0-9]+\.?[0-9]*')
     if [[ $nfo_file_size ]]; then
@@ -114,7 +125,7 @@ read_info_file() {
     generate_info_local
   fi
   debug_func 'info:exit'  #----debug---
-  unset img_url_com img_url_byr info_generated
 }
 
 #-------------------------------------#
+
